feat: 서비스 로직 인터페이스와 구현체로 분리
